+++ trace_stats/trace_stats.h
@@ -24,6 +24,144 @@
 namespace Dive
 {
 
+// Trace Statistics
+// Number of draw calls in BINNING, DIRECT, and TILED
+//  With LRZ enabled, and LRZ Write enabled
+//  With depth test enabled
+//  With depth write enabled
+//  With EarlyZ, LateZ enabled
+//  With CullMode enabled
+// Number of CLEARs  (???)
+// List of different unique viewports used
+// Min, Max, Total number of indices rendered
+// Number of tiles + size of tiles + location of tiles
+// Number of unique shaders, per stage (BINNING vs TILING)
+//  Min, Max, and total instruction counts
+//  Min, Max, and total GPR count
+// Number of CPEventWrites
+//  RESOLVE, FLUSH_COLOR, FLUSH_DEPTH, INVALIDATE_COLOR, INVALIDATE_DEPTH
+// Number of CpWaitForIdle()
+// Number of prefetches
+// Number of loads/stores to GMEM
+struct Stats
+{
+    enum Type : uint32_t
+    {
+        kNumBinnigPasses,
+        kNumTilingPasses,
+        kBinningDraws,
+        kDirectDraws,
+        kTiledDraws,
+        kDispatches,
+        kWaitMemWrites,
+        kWaitForIdle,
+        kWaitForMe,
+        kDepthTestEnabled,
+        kDepthWriteEnabled,
+        kEarlyZ,
+        kLateZ,
+        kEarlyLRZLateZ,
+        kLrzEnabled,
+        kLrzWriteEnabled,
+        kCullModeEnabled,
+        kTotalIndices,
+        kMinIndices,
+        kMaxIndices,
+        kMedianIndices,
+        kShaders,
+        kBinningVS,
+        kNonBinningVS,
+        kNonVS,
+        kTotalInstructions,
+        kMinInstructions,
+        kMaxInstructions,
+        kMedianInstructions,
+        kTotalGPRs,
+        kMinGPRs,
+        kMaxGPRs,
+        kMedianGPRs,
+        kTotalResolves,
+        kGmemToSysmemResolves,
+        kGmemToSysmemAndClearGmemResolves,
+        kClearGmemResolves,
+        kSysmemToGmemResolves,
+        kNumStats
+    };
+};
+
+constexpr std::array kStatMap = {
+    std::pair(Stats::kNumBinnigPasses, "Num Binning Passes"),
+    std::pair(Stats::kNumTilingPasses, "Num Tiling Passes"),
+    std::pair(Stats::kBinningDraws, "Num Draws (BINNING)"),
+    std::pair(Stats::kDirectDraws, "Num Draws (DIRECT)"),
+    std::pair(Stats::kTiledDraws, "Num Draws (TILED)"),
+    std::pair(Stats::kDispatches, "Num Dispatches"),
+    std::pair(Stats::kWaitMemWrites, "Num WaitMemWrites"),
+    std::pair(Stats::kWaitForIdle, "Num WaitForIdle"),
+    std::pair(Stats::kWaitForMe, "Num WaitForMe"),
+    std::pair(Stats::kDepthTestEnabled, "Num Draws with Depth Test Enabled"),
+    std::pair(Stats::kDepthWriteEnabled, "Num Draws with Depth Write Enabled"),
+    std::pair(Stats::kEarlyZ, "Num Draws with EarlyZ"),
+    std::pair(Stats::kLateZ, "Num Draws with LateZ"),
+    std::pair(Stats::kEarlyLRZLateZ, "Num Draws with Early LRZ & LateZ"),
+    std::pair(Stats::kLrzEnabled, "Num Draws with LRZ Enabled"),
+    std::pair(Stats::kLrzWriteEnabled, "Num Draws with LRZ Write Enabled"),
+    std::pair(Stats::kCullModeEnabled, "Num Draws with culling enabled"),
+    std::pair(Stats::kTotalIndices, "Total indices in all draws (includes non-indexed draws)"),
+    std::pair(Stats::kMinIndices, "\tMin indices in a single draw"),
+    std::pair(Stats::kMaxIndices, "\tMax indices in a single draw"),
+    std::pair(Stats::kMedianIndices, "\tMedian indices in a single draw"),
+    std::pair(Stats::kShaders, "Number of unique shaders"),
+    std::pair(Stats::kBinningVS, "\tNumber of BINNING VS"),
+    std::pair(Stats::kNonBinningVS, "\tNumber of non-BINNING VS"),
+    std::pair(Stats::kNonVS, "\tNumber of non-VS Shaders"),
+    std::pair(Stats::kTotalInstructions, "Total instructions in all shaders"),
+    std::pair(Stats::kMinInstructions, "\tMin instructions in a single shader"),
+    std::pair(Stats::kMaxInstructions, "\tMax instructions in a single shader"),
+    std::pair(Stats::kMedianInstructions, "\tMedian instructions in a single shader"),
+    std::pair(Stats::kTotalGPRs, "Total GPRs in all shaders"),
+    std::pair(Stats::kMinGPRs, "\tMin GPRs in a single shader"),
+    std::pair(Stats::kMaxGPRs, "\tMax GPRs in a single shader"),
+    std::pair(Stats::kMedianGPRs, "\tMedian GPRs in a single shader"),
+    std::pair(Stats::kTotalResolves, "Total resolves"),
+    std::pair(Stats::kGmemToSysmemResolves, "\tGmem to SysMem Resolves"),
+    std::pair(Stats::kGmemToSysmemAndClearGmemResolves, "\tGmem to SysMem Resolves and Clear Gmem"),
+    std::pair(Stats::kClearGmemResolves, "\tGmem Clears"),
+    std::pair(Stats::kSysmemToGmemResolves, "\tSysMem to Gmem Resolves"),
+};
+
+enum ViewPortStats
+{
+    kViewport,
+    kViewport_x,
+    kViewport_y,
+    kViewport_width,
+    kViewport_height,
+    kViewport_minDepth,
+    kViewport_maxDepth,
+    kNumViewportStats
+};
+
+inline const char *viewport_stats_desc[kNumViewportStats] = {
+    "Viewports", "x", "y", "width", "height", "minDepth", "maxDepth",
+};
+
+enum WindowScissorStats
+{
+    kWindowScissors,
+    kWindowScissors_tl_x,
+    kWindowScissors_br_x,
+    kWindowScissors_tl_y,
+    kWindowScissors_br_y,
+    kWindowScissors_Width,
+    kWindowScissors_Height,
+    kNumWindowScissorStats
+};
+
+inline const char *window_scissor_stats_desc[kNumWindowScissorStats] = {
+    "Window scissors", "tl_x", "br_x", "tl_y", "br_y", "Width", "Height"
+};
+
 struct Viewport
 {
     VkViewport m_vk_viewport;
